fazle-system/wbom/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
psycopg2-binary==2.9.10
redis==5.2.1
prometheus-fastapi-instrumentator==7.0.2
